+++ src/features/feature-discovery/index.js
@@ -1,0 +1,17 @@
+export {
+  trackFeatureSpotlightClick,
+  trackFeatureSpotlightDismissed,
+  trackFeatureSpotlightImpression,
+} from './model/featureDiscoveryAnalytics';
+export {
+  FEATURE_SPOTLIGHTS,
+  getFeatureSpotlight,
+  isSpotlightActive,
+  isSpotlightInWindow,
+  localYmd,
+  readSpotlightSeen,
+  spotlightStorageKey,
+  writeSpotlightSeen,
+} from './model/featureSpotlights';
+export { useFeatureSpotlight } from './model/useFeatureSpotlight';
+export { default as FeatureNewBadge } from './ui/FeatureNewBadge';